pyproject.toml
[project]
name = "omnivore"
version = "0.1.0"
description = "Quantitative trading platform for market prediction"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "psycopg[binary]>=3.1",
    "redis>=5.0",
    "rq>=1.16",
    "yfinance>=0.2",
    "pandas>=2.1",
    "numpy>=1.26",
    "pandas-ta>=0.3",
    "scikit-learn>=1.4",
    "xgboost>=2.0",
    "lightgbm>=4.3",
    "joblib>=1.3",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/omnivore"]

[tool.ruff]
line-length = 100
src = ["src"]
